refactor: extract estimate repository from route handler (Phase 3)
Moved the GET/PATCH/DELETE SQL logic from estimates/[id]/route.ts (729 lines)
into lib/estimates/repository.ts — getEstimateById, updateEstimateById,
deleteEstimateById. Route is now auth+validation+error handling only (155 lines).
